functions/requirements.txt
psycopg2-binary
google-generativeai
firebase-functions
python-telegram-bot
asyncio
